settle on citation handling
