+++ examples/example.html
@@ -235,6 +235,15 @@
                 cozy.control.title({ region: 'top.header.left' }).addTo(reader);
                 cozy.control.contents({ region: 'top.toolbar.left', skipLink: '.skip' }).addTo(reader);
 
+                cozy.control.citation({ 
+                    region: 'top.toolbar.left',
+                    citations: [
+                      { format: 'MLA', text: "Mock, Alex. <em>The Mock Life</em>. University Press, 2017." },
+                      { format: 'APA', text: "Mock, A. (2017). <em>The Mock Life</em>. Ann Arbor, MI: University Press." },
+                      { format: 'Chicago', text: "Mock, Alex. <em>The Mock Life</em>. Ann Arbor, MI: University Press, 2017." }
+                    ]
+                }).addTo(reader);
+
                 cozy.control.widget.panel({
                     region: 'top.toolbar.left',
                     template: '<div class="permalink-label"><label class="u-screenreader" for="permalink">Permalink</label><form><input data-slot="title" type="text" id="permalink" value="" readonly="readonly" onclick="this.select(); document.execCommand(\'copy\');"></form></div>',
